requirements.txt
# Function dependencies, for example:
# package==version
functions-framework==3.*
pymongo==4.6.1
boto3==1.34.16
cloudevents==1.9.0
importlib-metadata==6.7.0
click==8.1.7
cloudevents==1.9.0
deprecation==2.1.0
gunicorn==20.1.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
MarkupSafe==2.1.3
packaging==23.1
typing_extensions==4.7.1
watchdog==1.0.2
zipp==3.15.0
certifi==2023.11.17
python-dotenv==1.0.0
